# src/main/resources/static/api-docs/yaml/OnlineEventBooking-api-docs.yaml
openapi: 3.0.0
info:
  title: Online Event Booking API
  description: Rest API for Online Event Booking App
  version: 1.0.0
  contact:
    email: mukarramshk512@gmail.com
...
paths:
...
  /booking/save:
    post:
      tags:
        - EventBooking
      summary: Endpoints to save booking
      operationId: bookingSave
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventBookingModel'
      responses:
        '200':
          description: Successful event has been saved or updated
          content:
            text/plain:
              schema:
                type: string
  /booking/{venueId}:
    get:
      tags:
        - EventBooking
      summary: Get list of all Bookings By Specific VenueIdAndDate or ClientIdAndDate
      operationId: BookingList
      parameters:
        - in: query
          name: bookingId
          description: Unique reference of Booking record for client
          required: false
          schema:
            type: integer
            format: int64
        - in: path
          name: venueId
          description: Unique reference of Booking record for Venue
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful retrieval
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventBookingModel'
  /booking/delete/{id}:
    delete:
      tags:
        - EventBooking
      summary: Delete the Booking by id
      operationId: deleteBooking
      parameters:
        - in: path
          name: id
          description: Delete venue by id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful deleted
          content:
            text/plain:
              schema:
                type: string
...
components:
  schemas:
    EventBookingModel:
      type: object
      properties:
        id:
          type: integer
          format: int64
        date:
          type: string
          format: date
        time:
          type: string
        persons:
          type: integer
          format: int32
        totalAmount:
          type: number
          format: double
        balanceAmount:
          type: number
          format: double
        advanceAmount:
          type: number
          format: double
        menu:
          type: string
        extraDetails:
          type: string
        paymentMode:
          type: string
        isBookingConfirm:
          type: boolean
        client:
          type: object
          allOf:
            - $ref: '#/components/schemas/ClientModel'
        venue:
          type: object
          allOf:
            - $ref: '#/components/schemas/VenueModel'
    ClientModel:
      type: object
      properties:
        id:
          type: integer
          format: int64
        personInfo:
          type: object
          allOf:
            - $ref: '#/components/schemas/PersonInfoModel'
    VenueModel:
      type: object
      properties:
        id:
          type: integer
          format: int64
        perHeadRate:
          type: number
          format: double
        username:
          type: string
        password:
          type: string
        packageStartDate:
          type: string
          format: date
        packageEndsDate:
          type: string
          format: date
        isPackageActive:
          type: boolean
        venueInfo:
          type: object
          allOf:
            - $ref: '#/components/schemas/PersonInfoModel'
    PersonInfoModel:
      type: object
      properties:
        name:
          type: string
        contact:
          type: string
        city:
          type: string
        streetAddress:
          type: string
...